+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Publish edition-aware OpenAPI and compatibility contracts ([monica-openapi-edition-contracts](https://github.com/unbraind/monica-cli/blob/master/.agents/pm/issues/monica-openapi-edition-contracts.toon))
+
 ## 2026.7.26 - 2026-07-26
 
 ### Added